// apps/api/src/shared/persistence/prisma/models/Donor.ts
export type DonorWhereInput = {
  AND?: Prisma.DonorWhereInput | Prisma.DonorWhereInput[]
  OR?: Prisma.DonorWhereInput[]
  NOT?: Prisma.DonorWhereInput | Prisma.DonorWhereInput[]
  id?: Prisma.StringFilter<"Donor"> | string
  type?: Prisma.StringFilter<"Donor"> | string
  status?: Prisma.StringFilter<"Donor"> | string
  preferredAmount?: Prisma.DecimalNullableFilter<"Donor"> | runtime.Decimal | runtime.DecimalJsLike | number | string | null
  statusEndDate?: Prisma.DateTimeNullableFilter<"Donor"> | Date | string | null
  fullName?: Prisma.StringNullableFilter<"Donor"> | string | null
  email?: Prisma.StringNullableFilter<"Donor"> | string | null
  phoneCode?: Prisma.StringNullableFilter<"Donor"> | string | null
  phoneNumber?: Prisma.StringNullableFilter<"Donor"> | string | null
  userProfileId?: Prisma.StringNullableFilter<"Donor"> | string | null
  createdAt?: Prisma.DateTimeFilter<"Donor"> | Date | string
  updatedAt?: Prisma.DateTimeFilter<"Donor"> | Date | string
  version?: Prisma.IntFilter<"Donor"> | number
  deletedAt?: Prisma.DateTimeNullableFilter<"Donor"> | Date | string | null
  userProfile?: Prisma.XOR<Prisma.UserProfileNullableScalarRelationFilter, Prisma.UserProfileWhereInput> | null
  donations?: Prisma.DonationListRelationFilter
}

export type DonorOrderByWithRelationInput = {
  id?: Prisma.SortOrder
  type?: Prisma.SortOrder
  status?: Prisma.SortOrder
  preferredAmount?: Prisma.SortOrderInput | Prisma.SortOrder
  statusEndDate?: Prisma.SortOrderInput | Prisma.SortOrder
  fullName?: Prisma.SortOrderInput | Prisma.SortOrder
  email?: Prisma.SortOrderInput | Prisma.SortOrder
  phoneCode?: Prisma.SortOrderInput | Prisma.SortOrder
  phoneNumber?: Prisma.SortOrderInput | Prisma.SortOrder
  userProfileId?: Prisma.SortOrderInput | Prisma.SortOrder
  createdAt?: Prisma.SortOrder
  updatedAt?: Prisma.SortOrder
  version?: Prisma.SortOrder
  deletedAt?: Prisma.SortOrderInput | Prisma.SortOrder
  userProfile?: Prisma.UserProfileOrderByWithRelationInput
  donations?: Prisma.DonationOrderByRelationAggregateInput
}

export type DonorWhereUniqueInput = Prisma.AtLeast<{
  id?: string
  email?: string
  userProfileId?: string
  AND?: Prisma.DonorWhereInput | Prisma.DonorWhereInput[]
  OR?: Prisma.DonorWhereInput[]
  NOT?: Prisma.DonorWhereInput | Prisma.DonorWhereInput[]
  type?: Prisma.StringFilter<"Donor"> | string
  status?: Prisma.StringFilter<"Donor"> | string
  preferredAmount?: Prisma.DecimalNullableFilter<"Donor"> | runtime.Decimal | runtime.DecimalJsLike | number | string | null
  statusEndDate?: Prisma.DateTimeNullableFilter<"Donor"> | Date | string | null
  fullName?: Prisma.StringNullableFilter<"Donor"> | string | null
  phoneCode?: Prisma.StringNullableFilter<"Donor"> | string | null
  phoneNumber?: Prisma.StringNullableFilter<"Donor"> | string | null
  createdAt?: Prisma.DateTimeFilter<"Donor"> | Date | string
  updatedAt?: Prisma.DateTimeFilter<"Donor"> | Date | string
  version?: Prisma.IntFilter<"Donor"> | number
  deletedAt?: Prisma.DateTimeNullableFilter<"Donor"> | Date | string | null
  userProfile?: Prisma.XOR<Prisma.UserProfileNullableScalarRelationFilter, Prisma.UserProfileWhereInput> | null
  donations?: Prisma.DonationListRelationFilter
}, "id" | "email" | "userProfileId">
// ...
export type DonorUncheckedCreateInput = {
  id?: string
  type: string
  status: string
  preferredAmount?: runtime.Decimal | runtime.DecimalJsLike | number | string | null
  statusEndDate?: Date | string | null
  fullName?: string | null
  email?: string | null
  phoneCode?: string | null
  phoneNumber?: string | null
  userProfileId?: string | null
  createdAt?: Date | string
  updatedAt?: Date | string
  version?: number
  deletedAt?: Date | string | null
  donations?: Prisma.DonationUncheckedCreateNestedManyWithoutDonorInput
}
// ...
export type DonorUncheckedUpdateInput = {
  id?: Prisma.StringFieldUpdateOperationsInput | string
  type?: Prisma.StringFieldUpdateOperationsInput | string
  status?: Prisma.StringFieldUpdateOperationsInput | string
  preferredAmount?: Prisma.NullableDecimalFieldUpdateOperationsInput | runtime.Decimal | runtime.DecimalJsLike | number | string | null
  statusEndDate?: Prisma.NullableDateTimeFieldUpdateOperationsInput | Date | string | null
  fullName?: Prisma.NullableStringFieldUpdateOperationsInput | string | null
  email?: Prisma.NullableStringFieldUpdateOperationsInput | string | null
  phoneCode?: Prisma.NullableStringFieldUpdateOperationsInput | string | null
  phoneNumber?: Prisma.NullableStringFieldUpdateOperationsInput | string | null
  userProfileId?: Prisma.NullableStringFieldUpdateOperationsInput | string | null
  createdAt?: Prisma.DateTimeFieldUpdateOperationsInput | Date | string
  updatedAt?: Prisma.DateTimeFieldUpdateOperationsInput | Date | string
  version?: Prisma.IntFieldUpdateOperationsInput | number
  deletedAt?: Prisma.NullableDateTimeFieldUpdateOperationsInput | Date | string | null
  donations?: Prisma.DonationUncheckedUpdateManyWithoutDonorNestedInput
}